Filter Doctypes
Introduce two new function which helps to filter Doctypes which are already
included in another Doctype.

For example, if you have "sles/15-SP6/en-us" and "sles/*/en-us",
then the first would be already included in the latter. There is no need
to keep the first one (and maybe rebuild it and waste resources).