Fix #26: Missing checking <ref/> in XML config (#41)
In former times, the reference checking was done with the
`global-check-ref-list.xsl` stylesheet. Although this works, it
may be easier and can be customized further to implement it
in pure Python code.

This change allows to check the stitchfile against any <ref/>
which point to a non-existing target.

TODO: Make it more async?