Fix #118: Provide project Git config file (#120)
To create a reproducible build environment, we can't rely on the system
or user Git config. Depending on the options, it may let Git behave slightly
different than we expect.

For this reason, we need to ensure that we use project Git config that
is used whenever we call "git".